+++ RouteMeMobile/src/de/unipotsdam/nexplorer/client/android/callbacks/Loginable.java
@@ -2,5 +2,5 @@ package de.unipotsdam.nexplorer.client.android.callbacks;
 
 public interface Loginable {
 
-	public void loggedIn(int playerId);
+	public void loggedIn(long playerId);
 }
