+++ RouteMeMobile/src/de/unipotsdam/nexplorer/client/android/support/LoginObserver.java
@@ -2,10 +2,10 @@ package de.unipotsdam.nexplorer.client.android.support;
 
 import de.unipotsdam.nexplorer.client.android.callbacks.Loginable;
 
-public class LoginObserver extends ObserverWithParameter<Loginable, Integer> {
+public class LoginObserver extends ObserverWithParameter<Loginable, Long> {
 
 	@Override
-	protected void call(Loginable callback, Integer parameter) {
+	protected void call(Loginable callback, Long parameter) {
 		callback.loggedIn(parameter);
 	}
 }
